Move parsePublicKeys into new file
